+++ src/tightdb/objc/TDBTransaction.mm
@@ -63,10 +63,9 @@ using namespace std;
 -(TDBTable *)tableWithName:(NSString *)name
 {
     if ([name length] == 0) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:table_name_exception"
-                                                         reason:@"Name must be a non-empty NSString"
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:table_name_exception"
+                                       reason:@"Name must be a non-empty NSString"
+                                     userInfo:nil];
     }
     
     // If table does not exist in context, return nil
@@ -90,10 +89,9 @@ using namespace std;
 -(id)tableWithName:(NSString *)name asTableClass:(__unsafe_unretained Class)class_obj
 {
     if ([name length] == 0) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:table_name_exception"
-                                                         reason:@"Name must be a non-empty NSString"
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:table_name_exception"
+                                       reason:@"Name must be a non-empty NSString"
+                                     userInfo:nil];
     }
     
     // If table does not exist in context, return nil
@@ -136,24 +134,21 @@ using namespace std;
 -(TDBTable *)createTableWithName:(NSString*)name
 {
     if ([name length] == 0) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:table_name_exception"
-                                                         reason:@"Name must be a non-empty NSString"
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:table_name_exception"
+                                       reason:@"Name must be a non-empty NSString"
+                                     userInfo:nil];
     }
     
     if (m_read_only) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:core_read_only_exception"
-                                                         reason:@"Transaction is read-only."
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:core_read_only_exception"
+                                       reason:@"Transaction is read-only."
+                                     userInfo:nil];
     }
     
     if ([self hasTableWithName:name]) {
-        NSException* exception = [NSException exceptionWithName:@"tightdb:table_with_name_already_exists"
-                                                         reason:[NSString stringWithFormat:@"A table with the name '%@' already exists in the context.", name]
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:table_with_name_already_exists"
+                                       reason:[NSString stringWithFormat:@"A table with the name '%@' already exists in the context.", name]
+                                     userInfo:nil];
     }
 
     TDBTable* table = [[TDBTable alloc] _initRaw];
@@ -172,24 +167,21 @@ using namespace std;
 -(id)createTableWithName:(NSString*)name asTableClass:(__unsafe_unretained Class)class_obj
 {
     if ([name length] == 0) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:table_name_exception"
-                                                         reason:@"Name must be a non-empty NSString"
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+       @throw [NSException exceptionWithName:@"tightdb:table_name_exception"
+                                      reason:@"Name must be a non-empty NSString"
+                                    userInfo:nil];
     }
     
     if (m_read_only) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:core_read_only_exception"
-                                                         reason:@"Transaction is read-only."
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:core_read_only_exception"
+                                       reason:@"Transaction is read-only."
+                                     userInfo:nil];
     }
     
     if ([self hasTableWithName:name]) {
-        NSException* exception = [NSException exceptionWithName:@"tightdb:table_with_name_already_exists"
-                                                         reason:[NSString stringWithFormat:@"A table with the name '%@' already exists in the context.", name]
-                                                       userInfo:[NSMutableDictionary dictionary]];
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:table_with_name_already_exists"
+                                       reason:[NSString stringWithFormat:@"A table with the name '%@' already exists in the context.", name]
+                                     userInfo:nil];
     }
 
     TDBTable* table = [[class_obj alloc] _initRaw];
@@ -220,10 +212,9 @@ using namespace std;
         group->m_group = new tightdb::Group;
     }
     catch (std::exception& ex) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:core_exception"
-                                                         reason:[NSString stringWithUTF8String:ex.what()]
-                                                       userInfo:[NSMutableDictionary dictionary]];  // IMPORTANT: cannot not be nil !!
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:core_exception"
+                                       reason:[NSString stringWithUTF8String:ex.what()]
+                                     userInfo:nil];
     }
     group->m_is_owned  = YES;
     group->m_read_only = NO;
@@ -296,10 +287,9 @@ using namespace std;
         return nil;
     }
     catch (std::exception& ex) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:core_exception"
-                                                         reason:[NSString stringWithUTF8String:ex.what()]
-                                                       userInfo:[NSMutableDictionary dictionary]];  // IMPORTANT: cannot not be nil !!
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:core_exception"
+                                       reason:[NSString stringWithUTF8String:ex.what()]
+                                     userInfo:nil];
     }
     group->m_is_owned  = YES;
     group->m_read_only = NO;
@@ -350,10 +340,9 @@ using namespace std;
         return [[NSData alloc] initWithBytes:static_cast<const void *>(bd.data()) length:bd.size()];
     }
     catch (std::exception& ex) {
-        NSException *exception = [NSException exceptionWithName:@"tightdb:core_exception"
-                                                         reason:[NSString stringWithUTF8String:ex.what()]
-                                                       userInfo:[NSMutableDictionary dictionary]];  // IMPORTANT: cannot not be nil !!
-        [exception raise];
+        @throw [NSException exceptionWithName:@"tightdb:core_exception"
+                                       reason:[NSString stringWithUTF8String:ex.what()]
+                                     userInfo:nil];
     }
     return nil;
 }