+++ release_notes.md
@@ -37,10 +37,16 @@ Description......
 
 ### API breaking changes:
 
+* `findFirstRow` renamed to `indexOfFirstMatchingRow` on `TDBQuery`.
+* `findFirstRowFromIndex:` renamed to `indexOfFirstMatchingRowFromIndex:` on `TDBQuery`.
 * Return `NSNotFound` when appropriate.
+* `removeAllRows`, `removeRowAtIndex`, `removeLastRow`, `addRow` and `insertRow` methods 
+  on table now return void instead of BOOL.
+
 
 ### Enhancements:
 * `renameColumnWithIndex:to:` has been added to `TDBTable`.
+* `distinctValuesInColumnWithIndex` has been added to `TDBTable`.
 * `dateIsBetween::`, `doubleIsBetween::`, `floatIsBetween::` and `intIsBetween::`
    have been added to `TDBQuery`.
 * Column names can begin with non-capital letters too. The generated `addX`